First commit of this foobar branch,
To get this working, without changing the project files, I had to
disable an important looking assertion. Basically, in several hundred
thousand places the code was creating a delegate to fire off a RPC, but
when it came to send the rpc, the code had an assertion basically saying
"if this is a delegate, throw the assertion" but something must be
seriously wrong. It seems to work with the assertion disabled, not sure
about whether there is anything remaining of actual network
functionality though. =(